requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# Database
SQLAlchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
email-validator==2.3.0

# Security
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.10.1
cryptography==46.0.3

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.4
deep-translator==1.11.4
loguru==0.7.2
openpyxl==3.1.2

# Cloudinary
cloudinary==1.41.0

# Web Requests
requests==2.32.3
httpx==0.25.2
beautifulsoup4==4.14.3

# Date/Time
python-dateutil==2.8.2

# Type Checking
typing_extensions==4.12.2
